+++ src/js/api.ts
@@ -1,3 +1,5 @@
+/// <reference types="emscripten" />
+
 declare var Module: any;
 declare var Hiwire: any;
 declare var API: any;
@@ -103,7 +105,7 @@ export class PyodideAPI {
    * are available as members of ``FS.filesystems``:
    * ``IDBFS``, ``NODEFS``, ``PROXYFS``, ``WORKERFS``.
    */
-  static FS = {} as any;
+  static FS = {} as typeof FS;
   /**
    * An alias to the `Emscripten Path API
    * <https://github.com/emscripten-core/emscripten/blob/main/src/library_path.js>`_.